chore: remove schemas file
From now on I will try to make the changes to the files on the base
image instead of replacing it. This is so I don't need to pay attention
to every change they make on upstream, so the things that they change
upstream that are not things I personally want to do differently, are
also changed downstream.